requirements.txt
cryptography
requests
lxml
signxml
asn1crypto
endesive
python-dateutil
qrcode[pil]
pandas
openpyxl
customtkinter
Pillow
